# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Session:
    properties:
      chat_path:
        description: Path redirect for chat
        type: string
      chat_url:
        description: Chat/AI agents endpoint
        type: string
      container_name:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      is_active:
        type: boolean
      namespace:
        description: Uses project_uuid as namespace
        type: string
      preview_path:
        description: Path redirect for preview
        type: string
      preview_url:
        description: Service endpoints
        type: string
      project_id:
        type: integer
      project_uuid:
        description: UUID from another service
        type: string
      status:
        description: pending, running, stopped, error
        type: string
      token:
        type: string
      updated_at:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
      vscode_path:
        description: Path redirect for vscode
        type: string
      vscode_url:
        description: VS Code web endpoint
        type: string
    required:
    - project_id
    - project_uuid
    - user_id
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for PayPilot development session service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PayPilot Dev Session Service API
  version: "1.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: Check if the service is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Health check
      tags:
      - health
  /sessions:
    get:
      consumes:
      - application/json
      description: Get a list of all dev sessions with optional filtering
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by project ID
        in: query
        name: project_id
        type: integer
      - description: Filter by status (pending, running, stopped, error)
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List all dev sessions
      tags:
      - sessions
    post:
      consumes:
      - application/json
      description: Create a new dev session for a project in the no-code app generator
      parameters:
      - description: Session information
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/models.Session'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Session'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new development session
      tags:
      - sessions
  /sessions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a dev session by ID (also stops the associated container)
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a dev session
      tags:
      - sessions
    get:
      consumes:
      - application/json
      description: Get dev session details by session ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Session'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a dev session by ID
      tags:
      - sessions
  /sessions/project/{project_uuid}:
    get:
      consumes:
      - application/json
      description: Get an existing session for a project UUID, or create a new one
        if it doesn't exist
      parameters:
      - description: Project UUID
        in: path
        name: project_uuid
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        type: integer
      - description: Project ID
        in: query
        name: project_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Session'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get or create a dev session by project UUID
      tags:
      - sessions
swagger: "2.0"